General minor clean up. Fixes prediction issue with charger lights, and I think the IPC prototype is officially ready for action. CyborgComponent is dumb... So very dumb...
