+++ Content.Server/Nyanotrasen/Borgs/CyborgComponent.cs
@@ -1,5 +1,6 @@
 namespace Content.Server.Borgs
 {
+    [Obsolete("CyborgComponent is unused on Parkstation, see SiliconComponent")]
     [RegisterComponent]
     public sealed class CyborgComponent : Component
     {}
